Move RenderSVGResourceClipper into legacy https://bugs.webkit.org/show_bug.cgi?id=262114
Reviewed by Nikolas Zimmermann.

Move RenderSVGResourceClipper into legacy and rename it in order
to introduce LBSE RenderSVGResourceClipper variant, which will
offer stateless clipping for RenderLayer.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/rendering/ReferencedSVGResources.cpp:
(WebCore::ReferencedSVGResources::referencedClipperRenderer):
* Source/WebCore/rendering/ReferencedSVGResources.h:
* Source/WebCore/rendering/RenderBox.cpp:
(WebCore::RenderBox::hitTestClipPath const):
* Source/WebCore/rendering/RenderElement.cpp:
(WebCore::RenderElement::layerCreationAllowedForSubtree const):
* Source/WebCore/rendering/RenderLayer.cpp:
* Source/WebCore/rendering/svg/Legacy/LegacyRenderSVGResourceClipperInlines.h: Copied from Source/WebCore/rendering/svg/RenderSVGResourceClipperInlines.h.
(WebCore::LegacyRenderSVGResourceClipper::clipPathElement const):
(WebCore::LegacyRenderSVGResourceClipper::clipPathUnits const):
* Source/WebCore/rendering/svg/RenderSVGResource.cpp:
(WebCore::requestPaintingResource):
(WebCore::removeFromCacheAndInvalidateDependencies):
* Source/WebCore/rendering/svg/RenderSVGResourceSolidColor.cpp:
(WebCore::RenderSVGResourceSolidColor::applyResource):
* Source/WebCore/rendering/svg/SVGBoundingBoxComputation.cpp:
* Source/WebCore/rendering/svg/SVGRenderSupport.cpp:
(WebCore::SVGRenderSupport::intersectRepaintRectWithResources):
(WebCore::SVGRenderSupport::pointInClippingArea):
* Source/WebCore/rendering/svg/SVGRenderTreeAsText.cpp:
(WebCore::writeSVGResourceContainer):
(WebCore::writeResources):
* Source/WebCore/rendering/svg/SVGRenderingContext.cpp:
(WebCore::SVGRenderingContext::prepareToRenderSVGContent):
* Source/WebCore/rendering/svg/SVGResources.cpp:
(WebCore::SVGResources::buildCachedResources):
(WebCore::SVGResources::setClipper):
* Source/WebCore/rendering/svg/SVGResources.h:
(WebCore::SVGResources::clipper const):
* Source/WebCore/rendering/svg/SVGResourcesCycleSolver.cpp:
* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGResourceClipper.cpp: Renamed from Source/WebCore/rendering/svg/RenderSVGResourceClipper.cpp.
(WebCore::LegacyRenderSVGResourceClipper::LegacyRenderSVGResourceClipper):
(WebCore::LegacyRenderSVGResourceClipper::removeAllClientsFromCache):
(WebCore::LegacyRenderSVGResourceClipper::removeClientFromCache):
(WebCore::LegacyRenderSVGResourceClipper::applyResource):
(WebCore::LegacyRenderSVGResourceClipper::pathOnlyClipping):
(WebCore::LegacyRenderSVGResourceClipper::computeInputs):
(WebCore::LegacyRenderSVGResourceClipper::applyClippingToContext):
(WebCore::LegacyRenderSVGResourceClipper::drawContentIntoMaskImage):
(WebCore::LegacyRenderSVGResourceClipper::calculateClipContentRepaintRect):
(WebCore::LegacyRenderSVGResourceClipper::hitTestClipContent):
(WebCore::LegacyRenderSVGResourceClipper::resourceBoundingBox):
* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGResourceClipper.h: Renamed from Source/WebCore/rendering/svg/RenderSVGResourceClipper.h.
(WebCore::ClipperData::invalidate):
(isType):
* Source/WebCore/rendering/svg/legacy/LegacyRenderSVGResourceClipperInlines.h: Renamed from Source/WebCore/rendering/svg/RenderSVGResourceClipperInlines.h.
(WebCore::LegacyRenderSVGResourceClipper::clipPathElement const):
(WebCore::LegacyRenderSVGResourceClipper::clipPathUnits const):
* Source/WebCore/svg/SVGClipPathElement.cpp:
(WebCore::SVGClipPathElement::createElementRenderer):

Canonical link: https://commits.webkit.org/268500@main
